feat: Kerberos smart card module (#144)
This adds a new module for Kerberos smart cards.
This module can communicate with the real smart card using the WinSCard API to
sign some data.

You can notice that the code contains a lot of hardcoded values like input buffers
for the `SCardTransmit` function, control code for the `SCardControl` function, etc.
As of now, we do not have any APDU command builder or anything else. So, we are
forced to use hardcoded values. This will be refactored in a future patch.

The APDU command was extracted from data signing API calls recorded beforehand.

References:

* winscard.h header: https://learn.microsoft.com/en-us/windows/win32/api/winscard
* `pcsc` crate: https://docs.rs/pcsc/2.8.0