fix(ffi): use Mutex for the sspi methods syncronization.
According to [MSDN](https://learn.microsoft.com/en-us/windows/win32/secauthn/decryptmessage--general),
encrypt and decrypt message functions can be called from different threads simultaneously.
During the dev testing, we confirmed that these methods are called from the different threads.
But our implementation has no synchronization and we can not guarantee that it doesn't break
during the connection. To fix this, we wrapped SppiContext into a Mutex and implemented all
needed methods and traits for this wrapper.