fix(sspi): TLS buffers assigning during decryption.
Previously, we assumed that every buffer passed to the SSPI is a well-formed
TLS packet. But the caller can pass half of the TLS packet in the first
decrypt_message function call, and the other half in the second call.
So, we just set an empty buffer in the stream header and trailer.
During the testing, we verified that mstsc works well without them.