feat(ffi): winscard: PCSC-lite bindings (#246)
To support system-provided smart cards, we need to somehow access
them from `sspi-rs`. On Windows, we will use native WinSCard API.
We plan to use the `pcsc-lite` library on Linux and macOS; to do so,
we need these bindings.

_Why can't we just use the pcsc-sys crate?_

The main reason is that it links with the PCSC framework on macOS
which was deprecated for years.

_How the bindings work?_

The `sspi-rs` loads the pcsc-lite in the runtime, builds the functions
table, and returns it to the user. It allows us to:

* simplify build process;
* replace the library in the runtime.