fix(ffi): winscard: correctly implement SCardAccessStartedEvent function;
Previously, we returned hardcoded value from the `SCardAccessStartedEvent` function.
Unfortunately, it causes an error sometimes when using `sspi.dll` in `mstsc.exe`.
This problem was fixed by introducing one global event for the entire process.
We create one signaled event using Windows API and return a handle to it from
the `SCardAccessStartedEvent` function. In such a way we tell that our emulated
smart card is always ready. Moreover, this is a reason why we didn't use any
reference counters as the original API assumes.