feat(ffi): winscard: Winscard FFI improvements (#250)
_Implementation notes_

* Move the `clippy::undocumented_unsafe_blocks` lint to the `ffi/winscard` module
  lever instead of `ffi/winscard/system_scard`. Now all winscard-related code is checked.
* Refactor the code and remove all panics.
* Improve the memory management.

_Memory management improvements_

Many winscard functions can treat the resulting data differently based on the input parameters.

* When the input data pointer is `NULL`, we should write out data len in the input data
  len pointer.
* When the input data length pointer value is `SCARD_AUTOALLOCATE`, we should allocate
  memory by ourselves, cast the output data pointer to pointer to pointer to memory, and
  write the memory address.
* Otherwise, we copy the data and data length into corresponding pointers.

And there are a lot of functions with such behavior. To simplify this process and reduce
many `if`s/`match`s, We've introduced the `RequestBufferType` and `OutBuff` types.
They are well documented, so you can read its description in the code.

Moreover, we put the effort into keeping all memory management inside the `WinScardContexthandle`
structure implementation. Now it's easier to manage the memory and search for problems if there are any.