feat: add `query_context_session_key`
This addition is an implementation for the SSP API `QueryContextAttributesEx(SECPKG_ATTR_SESSION_KEY)`.
It is required for protocols such as SMB, and adding it to the `Sspi` trait, enables access to it across
all the SSP packages implemented.

This adds the option for using Kerberos and Negotiate session keys.