Bugfix: Kerberos tokens broken since upgrade to picky-krb 0.10.0
Since the upgrade to the newer version of picky-krb, the encoded kerberos neg token had an additional sequence TLV wrapping around it's value (because KrbMessage's encoding is now wrapped with a `Container` adding sequence TLV to it (the data is the same inside the sequence).
Using `ApplicationTag0` allows, as documented, skipping the sequence TLV, and encoding the value directly in the application tag.