feat(ntlm): Add the method to set the channel bindings for a session
NTLM authentication exchange over a TLS connection can provide a channel
binding token (calculated from the server's certificate by the procedure
described in RFC 5929) to prevent relaying attacks. The token can be
set as soon as the `Ntlm` struct is initialized, and will be transmitted
(appropriately transformed) to the server in the AUTHENTICATE message
of the session.

See also Section 3.1.5.2.1 (Client Receives a CHALLENGE_MESSAGE) of the
NTLM specification.

Closes #240